Updated memoryHandler

+++ TS/src/Utils/MemoryHandler.ts
@@ -4,15 +4,11 @@ export type IDestroyable = { __destroy: (recursive?: boolean) => void }
 
 const initMemoryHandler = () => {
     let numCreatedObjects = 0
-    let numCreatedArrays = 0
 
     const debugObjects: { [x: string]: number } = {}
-    const debugArrays: { [x: string]: number } = {}
-
     const cachedObjects: any[] = []
-    const cachedArrays: any[] = []
 
-    const purgeObject = (obj: any, recursive?: boolean) => {
+    const purgeObject = (obj: any | any[], recursive?: boolean) => {
         for (const [k] of pairs(obj)) {
             recursive && typeof obj[k] === 'object' && obj[k].__destroy?.(obj[k], recursive)
             obj[k] = undefined
@@ -34,36 +30,14 @@ const initMemoryHandler = () => {
         }
     }
 
-    const purgeArray = (arr: any[], recursive?: boolean) => {
-        for (const [k] of pairs(arr)) {
-            recursive && typeof arr[k] === 'object' && arr[k].__destroy?.(arr[k], recursive)
-            ;(arr as any)[k] = undefined
-        }
-
-        const meta = getmetatable(arr) as any
-
-        if (meta.__debugName) {
-            if (meta.__debugName && debugArrays[meta.__debugName]) {
-                debugArrays[meta.__debugName]--
-
-                if (debugArrays[meta.__debugName] === 0) {
-                    ;(debugArrays[meta.__debugName] as any) = undefined
-                }
-            }
-
-            meta.__debugName = undefined
-            meta.__destroyed = true
-        }
-    }
-
     const destroyObject = (self: any, recursive = false) => {
+        if (!self.__destroy) {
+            print(info().GetStackTrace())
+            throw 'Object is not memory handled'
+        }
+
         purgeObject(self, recursive)
         arrayPush(cachedObjects, self)
-    }
-
-    const destroyArray = (self: any[], recursive = false) => {
-        purgeArray(self, recursive)
-        arrayPush(cachedArrays, self)
     }
 
     const getObjectMeta = (debugName?: string) => {
@@ -96,38 +70,7 @@ const initMemoryHandler = () => {
         return meta
     }
 
-    const getArrayMeta = (debugName?: string) => {
-        const meta: any = {
-            __gc: (self: any[]) => {
-                purgeArray(self)
-            },
-            __newindex: (self: any[], k: any, v: any) => {
-                if (meta.__destroyed) {
-                    print(info().GetStackTrace())
-                    throw 'Writing a destroyed array'
-                }
-
-                rawset(self, k, v)
-            },
-            __index: (_self: any[], key: string) => {
-                if (meta.__destroyed) {
-                    print(info().GetStackTrace())
-                    throw 'Reading a destroyed array'
-                }
-
-                if (key === '__destroy') {
-                    return destroyArray
-                }
-            },
-        }
-
-        debugName && (meta['__debugName'] = debugName)
-
-        return meta
-    }
-
     const defaultObjectMeta = getObjectMeta()
-    const defaultArrayMeta = getArrayMeta()
 
     type ITarget = { debugName: string | number; count: number }
     const targetCompare = (a: ITarget, b: ITarget) => b.count < a.count
@@ -149,7 +92,7 @@ const initMemoryHandler = () => {
             let i = 0
 
             for (const s of sortedTargets) {
-                if (i++ < 10) {
+                if (i++ < 8) {
                     d += (d.length > 0 ? ', ' : '') + `${s.debugName}: ${s.count}`
                 }
             }
@@ -160,59 +103,83 @@ const initMemoryHandler = () => {
         sortedTargets.__destroy(true)
     }
 
+    const getEmptyClass = <T extends new (...args: any) => any>(
+        classInstance: T,
+        ..._params: ConstructorParameters<T>
+    ) => {
+        let params = _params as unknown[] // We cast params to array so that TSTL adds a +1 since arrays are 1-indexed
+        let debugName: string | undefined = undefined
+
+        if (typeof classInstance === 'string') {
+            debugName = classInstance
+            classInstance = params[0] as T
+            params = params.slice(1)
+        }
+
+        const obj = getEmptyObject<InstanceType<T>>(debugName)
+
+        // local function __TS__Class(self)
+        //     local c = {prototype = {}}
+        //     c.prototype.__index = c.prototype
+        //     c.prototype.constructor = c
+        //     return c
+        // end
+
+        // local function __TS__New(target, ...)
+        //     local instance = setmetatable({}, target.prototype)
+        //     instance:____constructor(...)
+        //     return instance
+        // end
+
+        // MIGHT NEED IN FUTURE
+        // obj.__index = classInstance.prototype
+
+        // MIGHT NEED IN FUTURE; basically do new classInstance above somewhere and cache one per name
+        // for (const [k, v] of pairs(classCache[classInstance.name])) {
+        //     if (typeof k !== 'string') continue
+        //     if (k.startsWith('_')) continue
+        //     obj[k] = v
+        // }
+
+        classInstance.prototype.____constructor?.(obj, ...params)
+
+        return obj
+    }
+
+    const getEmptyObject = <T>(debugName?: string) => {
+        let obj: T & IDestroyable = cachedObjects.shift()
+
+        if (!!obj) {
+            // Causes bugs if debugName changes where getEmptyObject gets called
+            if (debugName) {
+                ;(getmetatable(obj) as any).__debugName = debugName
+                ;(getmetatable(obj) as any).__destroyed = false
+            }
+        } else {
+            obj = {} as any
+            numCreatedObjects++
+            setmetatable(obj, debugName ? getObjectMeta(debugName) : defaultObjectMeta)
+        }
+
+        if (debugName) {
+            if (!debugObjects[debugName]) {
+                debugObjects[debugName] = 0
+            }
+
+            debugObjects[debugName]++
+        }
+
+        return obj
+    }
+
     return {
-        getEmptyObject: <T>(debugName?: string) => {
-            let obj: T & IDestroyable = cachedObjects.shift()
-
-            if (obj) {
-                // Causes bugs if debugName changes where getEmptyObject gets called
-                if (debugName) {
-                    ;(getmetatable(obj) as any).__debugName = debugName
-                    ;(getmetatable(obj) as any).__destroyed = false
-                }
-            } else {
-                obj = {} as any
-                numCreatedObjects++
-                setmetatable(obj, debugName ? getObjectMeta(debugName) : defaultObjectMeta)
-            }
-
-            if (debugName) {
-                if (!debugObjects[debugName]) {
-                    debugObjects[debugName] = 0
-                }
-
-                debugObjects[debugName]++
-            }
-
-            return obj
-        },
+        getEmptyClass,
+        getEmptyObject,
         getEmptyArray: <T>(debugName?: string) => {
-            let arr: T[] & IDestroyable = cachedArrays.shift()
-
-            if (arr) {
-                // Causes bugs if debugName changes where getEmptyArray gets called
-                if (debugName) {
-                    ;(getmetatable(arr) as any).__debugName = debugName
-                    ;(getmetatable(arr) as any).__destroyed = false
-                }
-            } else {
-                arr = [] as any
-                numCreatedArrays++
-                setmetatable(arr, debugName ? getArrayMeta(debugName) : defaultArrayMeta)
-            }
-
-            if (debugName) {
-                if (!debugArrays[debugName]) {
-                    debugArrays[debugName] = 0
-                }
-
-                debugArrays[debugName]++
-            }
-
-            return arr
+            return getEmptyObject<T[]>(debugName) as T[] & IDestroyable
         },
         destroyObject,
-        destroyArray,
+        destroyArray: destroyObject,
         cloneArray: <T>(arr: T[]) => {
             const newArray = MemoryHandler.getEmptyArray<T>()
 
@@ -225,10 +192,8 @@ const initMemoryHandler = () => {
         printDebugInfo: () => {
             print('MemoryHandler')
             print(`Objects: ${numCreatedObjects - cachedObjects.length}/${numCreatedObjects}`)
-            print(`Arrays: ${numCreatedArrays - cachedArrays.length}/${numCreatedArrays}`)
 
             printDebugNames('objects', debugObjects)
-            printDebugNames('arrays', debugArrays)
 
             if ((_G as any)['trackPrintMap']) {
                 printDebugNames('globals', (_G as any)['__fakePrintMap'])
